WIP: Send test result to TO
Grabbing test result via a hookimplementation and sending it to TO after test is done but before teardown.

One thing that still needs to be done is clean the code up.
Because fixtures can't access class objects, many things are unavailable.

Solution to this is create a class fixture and use that on the function fixture